tests/wpt/tests/css/css-backgrounds/parsing/background-origin-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders: getComputedStyle().backgroundOrigin</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-origin">
<meta name="assert" content="background-origin computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("background-origin", "border-box");
test_computed_value("background-origin", "padding-box");
test_computed_value("background-origin", "content-box");
test_computed_value("background-origin", "border-box, padding-box, content-box");

document.getElementById("target").style['background-image'] = "none, none";
test_computed_value("background-origin", "border-box", "border-box", "multiple base values");
test_computed_value("background-origin", "border-box, padding-box", "border-box, padding-box", "multiple base values");
test_computed_value("background-origin", "border-box, padding-box, content-box", "border-box, padding-box, content-box", "multiple base values");
</script>
</body>
</html>
